DELIA-40553: Duplicate timedMetadata reported and wrong time value
Reason for change: During linear playlist refresh , already reported Metadata
was getting reported .Also intermittenly time of Metadata was showing outside
culling window. Fixed the duplicate handling across playlist refresh.On underflow
retune, culled value was resetted to 0 causing timed Metadata time to reset.
Test Procedure: Tested with Peacock linear content and checked Metadata
Risks: Low

(cherry picked from commit d147f3cfb18520f68d5576d22ca0823257d56193)

+++ jsbindings/jsutils.h
@@ -79,6 +79,6 @@ void aamp_dispatchEventToJS(JSContextRef context, JSObjectRef callback, JSObject
 AAMPEventType aampPlayer_getEventTypeFromName(const char* szName);
 const char* aampPlayer_getNameFromEventType(AAMPEventType type);
 
-JSObjectRef aamp_CreateTimedMetadataJSObject(JSContextRef context, double timeMS, const char* szName, const char* szContent, const char* id, double durationMS);
+JSObjectRef aamp_CreateTimedMetadataJSObject(JSContextRef context, long long timeMS, const char* szName, const char* szContent, const char* id, double durationMS);
 
 #endif /* __AAMP_JSUTILS_H__ */
